Split agent behavior tests
